Fix: Extract and apply authentication middleware centrally
- I created a dedicated authentication middleware in backend/middleware/auth.js.
- I updated backend/server.js to import and use this central middleware for protected routes.
- I removed the inline authentication logic from server.js.
- I cleaned up individual route files by removing any local imports of the authentication middleware, as it's now applied globally.
