+++ backend/server.js
@@ -1,8 +1,9 @@
 const express = require('express');
 const path = require('path');
 require('dotenv').config();
-// Specific import as requested
-const { User, sequelize } = require('./models');
+// User model removed, sequelize kept. Auth middleware imported.
+const { sequelize } = require('./models');
+const { authenticateAndAttachUser } = require('./middleware/auth');
 
 // Corrected path to existing payroll routes file
 const payrollRoutes = require('./routes/payroll');
@@ -11,32 +12,6 @@ const app = express();
 
 // Init Middleware
 app.use(express.json({ extended: false }));
-
-// FAKE Authentication Middleware (re-created based on previous observation of such a middleware)
-// This is a placeholder and would need a proper implementation
-const authenticateAndAttachUser = async (req, res, next) => {
-    try {
-        // This is a mock user fetch. In a real app, this would involve token validation etc.
-        // Using a known email from potential seed data or test cases.
-        const loggedInUser = await User.findOne({
-            where: { email: 'manager.rh@techsolutions.ma' } // Example email
-        });
-
-        if (!loggedInUser) {
-            // If no user found, for testing, we might allow routes to proceed without req.user
-            // or return an error, depending on strictness. For now, proceed.
-            console.warn('Mock Authentication: User not found for testing with email manager.rh@techsolutions.ma. Proceeding without req.user.');
-            return next();
-        }
-        req.user = loggedInUser;
-        next();
-    } catch (error) {
-        console.error("Mock auth error:", error);
-        // Even with an error, proceed so app doesn't hang, but log it.
-        // In a real app, might return 500.
-        next();
-    }
-};
 
 // Define Routes
 // Other routes like auth, employees, dependents are removed as per the focus on payrollRoutes and new structure.
